requirements.txt
# Core Dependencies
scapy>=2.5.0
PyQt6>=6.6.0
pyqtgraph>=0.13.3
numpy>=1.24.0
pandas>=2.0.0

# Async & Performance
asyncio-mqtt>=0.16.1
aiofiles>=23.2.1
redis>=5.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Export & Reporting
reportlab>=4.0.0
openpyxl>=3.1.0
jinja2>=3.1.2

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Security & Crypto
cryptography>=41.0.0
pyopenssl>=23.3.0

# Network Tools
dnspython>=2.4.0
netifaces>=0.11.0
psutil>=5.9.0

# Utilities
python-dateutil>=2.8.2
colorama>=0.4.6
rich>=13.6.0
pyyaml>=6.0.1
